Pundit version

+++ Backend/app/controllers/room_bookings_controller.rb
@@ -2,41 +2,41 @@
 
 class RoomBookingsController < ApplicationController
   before_action :set_room_booking, only: %i[show update destroy]
-  # before_action :set_user, only: %i[create]
 
   # GET /room_bookings
   def index
     @room_bookings = RoomBooking.all
+    authorize @room_bookings
     render json: SerializerHelper::serialize(:RoomBookingSerializer, @room_bookings)
   end
 
   # GET /room_bookings/1
   def show
+    authorize @room_booking
     render json: SerializerHelper::serialize(:RoomBookingSerializer, @room_booking)
   end
 
   # POST /room_bookings
   def create
     @room_booking = current_user.room_bookings.new(room_booking_params)
-
+    authorize @room_bookings
+    @room = Room.where(:id => params[:room_id]).first
+    @room_booking.available_seats = @room.seats_count
     if @room_booking.save
       render json: SerializerHelper::serialize(:RoomBookingSerializer, @room_booking), status: :created, location: @room_booking
 
-    # else
-    #   render json: @room_booking.errors, status: :unprocessable_entity
+    else
+      render json: @room_booking.errors, status: :unprocessable_entity
     end
   end
 
   # DELETE /room_bookings/1
   def destroy
+    authorize @room_booking
     current_user.room_booking.destroy
   end
 
   private
-
-  # def set_user
-  #   @user = current_user
-  # end
 
   # Use callbacks to share common setup or constraints between actions.
   def set_room_booking
@@ -45,6 +45,6 @@ class RoomBookingsController < ApplicationController
 
   # Only allow a list of trusted parameters through.
   def room_booking_params
-    params.require(:room_booking).permit(:available_seats, :starts_at, :ends_at, :room_id)
+    params.require(:room_booking).permit( :starts_at, :ends_at, :room_id)
   end
 end
